chatbot/requirements.txt
python-dotenv
langchain
langchain-huggingface
langchain-community
transformers
faiss-cpu
python-dateutil
pypdf
langchain-unstructured
unstructured[pdf]
python-magic-bin